equals() and hashCode() moved to parent class: QueryResult.java
