+++ query/src/main/java/io/keen/client/java/result/DoubleResult.java
@@ -1,10 +1,5 @@
 package io.keen.client.java.result;
 
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
 /**
  * DoubleResult is for if the QueryResult object is of type Double.
  *
@@ -35,20 +30,5 @@ public class DoubleResult extends QueryResult {
     @Override
     public double doubleValue() {
         return result;
-    }
-
-    @Override
-    public String toString() {
-        return ReflectionToStringBuilder.toString(this, ToStringStyle.SHORT_PREFIX_STYLE);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return EqualsBuilder.reflectionEquals(this, obj);
-    }
-
-    @Override
-    public int hashCode() {
-        return HashCodeBuilder.reflectionHashCode(this);
     }
 }
