+++ query/src/main/java/io/keen/client/java/result/QueryResult.java
@@ -1,5 +1,10 @@
 package io.keen.client.java.result;
 
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
+
 import java.util.List;
 import java.util.Map;
 
@@ -93,4 +98,19 @@ public abstract class QueryResult {
      * @return map of Group to QueryResult's, which is IllegalStateException in abstract class.
      */
     public Map<Group, QueryResult> getGroupResults() { throw new IllegalStateException(); }
+
+    @Override
+    public String toString() {
+        return ReflectionToStringBuilder.toString(this, ToStringStyle.SHORT_PREFIX_STYLE);
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        return EqualsBuilder.reflectionEquals(this, obj);
+    }
+
+    @Override
+    public int hashCode() {
+        return HashCodeBuilder.reflectionHashCode(this);
+    }
 }
